fix: replace int with long for the quantities

+++ src/main/java/com/ramanbabich/investment/limetraderclient/account/model/AccountBalance.java
@@ -27,7 +27,7 @@ public record AccountBalance(
     String marginType,
     String restriction,
     String restrictionReason,
-    Integer daytradesCount,
+    Long daytradesCount,
     BigDecimal accountValueTotal,
     BigDecimal cash,
     Integer optionLevel,
